added CencelTrade method, started to add users logic

+++ trade_service/internal/services/models/users.go
@@ -15,7 +15,7 @@ type (
 
 	Tariff struct {
 		ID         uint
-		Comission  uint
+		Comission  uint8
 		Permission uint
 	}
 )
